No longer interface file, all in py_multipatch
